Reorganize windowed depth op
Collapse the `windows` and `beddepth` modules into a single
`window_depth` module, which is where we do everything relevant to this
command. I have also renamed stuff to `window-depth` instead of
`bed-depth` because I think this better describes what we're doing here,
in terms of the end-to-end purpose.

This also removes the "example" program I added in #252 because this
fully supersedes it.
