+++ flatgfa/src/cli/main.rs
@@ -51,7 +51,7 @@ enum Command {
     SeqExport(cmds::SeqExport),
     SeqImport(cmds::SeqImport),
     PangenotypeMatrix(cmds::PangenotypeMatrix),
-    BedDepth(cmds::BedDepth),
+    WindowDepth(cmds::WindowDepth),
 }
 
 fn main() -> Result<(), &'static str> {
@@ -175,8 +175,8 @@ fn main() -> Result<(), &'static str> {
         Some(Command::PangenotypeMatrix(sub_args)) => {
             cmds::pangenotype_matrix(&gfa, sub_args);
         }
-        Some(Command::BedDepth(sub_args)) => {
-            cmds::bed_depth(&gfa, sub_args);
+        Some(Command::WindowDepth(sub_args)) => {
+            cmds::window_depth(&gfa, sub_args);
         }
         None => {
             // Just emit the GFA or FlatGFA file.
