+++ flatgfa/src/ops/mod.rs
@@ -1,4 +1,3 @@
-pub mod beddepth;
 pub mod bench;
 pub mod chop;
 pub mod depth;
@@ -6,4 +5,4 @@ pub mod extract;
 pub mod gaf;
 pub mod pangenotype;
 pub mod position;
-pub mod windows;
+pub mod window_depth;
